Added an optional publish.filter configuration item.
The publish.filter allows additional criteria to be specified for services to be publised in the zookeeper discovery server. It allows you to publish a subset of the remoted services.
The publish.filter is an optional configuration item. Not specifying the item means that no extra filter is in place. It is specified through the OSGi Config Admin service along with the zookeeper.port, zookeeper.host etc items. The following example only publishes endpoints that are created on port 80 to discovery:
  publish.filter = (endpoint.id=*:80/*)